+++ python/GafferVDBTest/__init__.py
@@ -42,6 +42,7 @@ from .PointsGridToPointsTest import PointsGridToPointsTest
 from .ModuleTest import ModuleTest
 from .SphereLevelSetTest import SphereLevelSetTest
 from .PointsToLevelSetTest import PointsToLevelSetTest
+from .VolumeScatterTest import VolumeScatterTest
 
 if __name__ == "__main__":
 	import unittest